refactor(backup): delete LegacyBackupManager
Its last job was producing a LAN-transfer archive in the v1 layout — a
`data.json` beside an empty `Data/` — which nothing in v2 can restore: admission
expects `manifest.json` + `backup.sqlite` + `resources/`. Its renderer payload
producer was the matching v1 shape and went with the rest of the renderer's
backup service. Keeping the generator would have preserved a pipeline whose
output no version of this app can read; the relaunch needs the v2 archive
`BackupService.export()` already makes.

The one piece worth keeping moves to the module that owns it:
`deleteTransferFile` on `LanTransferService`, with its path-traversal tests
rewritten against the new home — the path still arrives from the renderer, so
the guard still matters.

`Backup_*` IPC channels are now zero, and the backup preload namespace with
them. Everything backup-related crosses the boundary through IpcApi.
