--- src/main/services/__tests__/LegacyBackupManager.test.ts
@@ -1,383 +1,0 @@
-import type * as CryptoModule from 'node:crypto'
-
-import type * as PathModule from 'path'
-import { beforeEach, describe, expect, it, vi } from 'vitest'
-
-// Mock path module to normalize all paths to POSIX format for cross-platform consistency
-// This ensures path operations work the same way regardless of the actual OS
-vi.mock('path', async () => {
-  const actual: typeof PathModule = await vi.importActual('path')
-  return {
-    ...actual,
-    sep: '/', // Always use forward slash for consistency
-    delimiter: ':',
-    join: (...args: string[]) => {
-      // Join with forward slashes, normalizing away backslashes
-      return actual.join(...args).replace(/\\/g, '/')
-    },
-    normalize: (p: string) => {
-      // Normalize path separators and remove redundant slashes
-      return actual.normalize(p).replace(/\\/g, '/')
-    },
-    resolve: (...args: string[]) => {
-      // For paths starting with / (Unix-style), use posix.resolve to avoid drive letter prefix
-      if (args.some((arg) => typeof arg === 'string' && arg.startsWith('/'))) {
-        return actual.posix.resolve(...args.map((a) => String(a).replace(/\\/g, '/')))
-      }
-      // For relative or Windows paths, use native resolve
-      return actual.resolve(...args).replace(/\\/g, '/')
-    },
-    isAbsolute: (p: string) => actual.isAbsolute(p) || String(p).startsWith('/'),
-    dirname: (p: string) => actual.dirname(p).replace(/\\/g, '/'),
-    basename: actual.basename,
-    extname: actual.extname,
-    relative: (from: string, to: string) =>
-      actual.relative(from.replace(/\\/g, '/'), to.replace(/\\/g, '/')).replace(/\\/g, '/'),
-    // Keep native POSIX and win32 for direct use if needed
-    posix: actual.posix,
-    win32: actual.win32
-  }
-})
-
-// Use vi.hoisted to define mocks that are available during hoisting
-const { mockLogger, mockBackupService, mockWindowManager, mockRandomUUID } = vi.hoisted(() => {
-  return {
-    mockLogger: {
-      debug: vi.fn(),
-      info: vi.fn(),
-      warn: vi.fn(),
-      error: vi.fn()
-    },
-    mockBackupService: {
-      export: vi.fn(),
-      prepareRestore: vi.fn(),
-      armRestore: vi.fn()
-    },
-    mockWindowManager: { broadcastToType: vi.fn(), getWindowsByType: vi.fn(() => []) },
-    mockRandomUUID: vi.fn()
-  }
-})
-
-vi.mock('node:crypto', async () => {
-  const actual = await vi.importActual<typeof CryptoModule>('node:crypto')
-  return { ...actual, randomUUID: mockRandomUUID }
-})
-
-vi.mock('@logger', () => ({
-  loggerService: {
-    withContext: () => mockLogger
-  }
-}))
-
-vi.mock('electron', () => ({
-  app: {
-    getPath: vi.fn((key: string) => {
-      if (key === 'temp') return '/tmp'
-      if (key === 'userData') return '/mock/userData'
-      return '/mock/unknown'
-    }),
-    getVersion: vi.fn(() => '2.0.0')
-  }
-}))
-
-vi.mock('fs-extra', () => ({
-  default: {
-    pathExists: vi.fn(),
-    remove: vi.fn(),
-    rename: vi.fn(),
-    ensureDir: vi.fn(),
-    emptyDir: vi.fn(),
-    copy: vi.fn(),
-    readdir: vi.fn(),
-    lstat: vi.fn(),
-    stat: vi.fn(),
-    realpath: vi.fn(),
-    readFile: vi.fn(),
-    writeFile: vi.fn(),
-    readJson: vi.fn(),
-    writeJson: vi.fn(),
-    createWriteStream: vi.fn(),
-    createReadStream: vi.fn(),
-    lstatSync: vi.fn(),
-    readdirSync: vi.fn(),
-    openSync: vi.fn(),
-    fsyncSync: vi.fn(),
-    closeSync: vi.fn(),
-    existsSync: vi.fn(),
-    promises: {
-      mkdir: vi.fn(),
-      readFile: vi.fn()
-    }
-  },
-  pathExists: vi.fn(),
-  remove: vi.fn(),
-  rename: vi.fn(),
-  ensureDir: vi.fn(),
-  emptyDir: vi.fn(),
-  copy: vi.fn(),
-  readdir: vi.fn(),
-  lstat: vi.fn(),
-  stat: vi.fn(),
-  realpath: vi.fn(),
-  readFile: vi.fn(),
-  writeFile: vi.fn(),
-  readJson: vi.fn(),
-  writeJson: vi.fn(),
-  createWriteStream: vi.fn(),
-  createReadStream: vi.fn(),
-  lstatSync: vi.fn(),
-  readdirSync: vi.fn(),
-  openSync: vi.fn(),
-  fsyncSync: vi.fn(),
-  closeSync: vi.fn(),
-  existsSync: vi.fn(),
-  promises: {
-    mkdir: vi.fn(),
-    readFile: vi.fn()
-  }
-}))
-
-vi.mock('@application', () => ({
-  application: {
-    get: vi.fn((name: string) => {
-      if (name === 'MainWindowService') {
-        return { getMainWindow: vi.fn() }
-      }
-      if (name === 'WindowManager') {
-        return mockWindowManager
-      }
-      if (name === 'BackupService') {
-        return mockBackupService
-      }
-      throw new Error(`[MockApplication] Unknown service: ${name}`)
-    }),
-    getPath: vi.fn((key: string, filename?: string) => {
-      const paths: Record<string, string> = {
-        'app.userdata': '/mock/userData',
-        'app.userdata.data': '/mock/userData/Data',
-        'app.database.file': '/mock/userData/Data/cherrystudio.sqlite',
-        'feature.backup.temp': '/mock/temp/backup',
-        'feature.backup.restore.file': '/mock/userData/Data/restore-journal.json',
-        'feature.backup.restore.staging': '/mock/userData/restore-staging',
-        'feature.agents.claude.root': '/mock/userData/Data/Agents/.claude',
-        'feature.lan_transfer.temp': '/tmp/cherry-studio/lan-transfer'
-      }
-      const base = paths[key] ?? '/mock/unknown'
-      return filename ? `${base}/${filename}` : base
-    })
-  }
-}))
-
-vi.mock('../WebDav', () => ({
-  default: vi.fn()
-}))
-
-vi.mock('../S3Storage', () => ({
-  default: vi.fn()
-}))
-
-vi.mock('archiver', () => ({
-  ZipArchive: vi.fn()
-}))
-
-// Import after mocks
-import * as fs from 'fs-extra'
-import * as path from 'path'
-
-import BackupManager from '../LegacyBackupManager'
-
-// Helper to construct platform-independent paths for assertions
-// The implementation uses path.normalize() which converts to platform separators
-const normalizePath = (p: string): string => path.normalize(p)
-
-describe('BackupManager.deleteLanTransferBackup - Security Tests', () => {
-  let backupManager: BackupManager
-
-  beforeEach(() => {
-    vi.clearAllMocks()
-    backupManager = new BackupManager()
-  })
-
-  describe('Normal Operations', () => {
-    it('should delete valid file in allowed directory', async () => {
-      vi.mocked(fs.pathExists).mockResolvedValue(true as never)
-      vi.mocked(fs.remove).mockResolvedValue(undefined as never)
-
-      const validPath = '/tmp/cherry-studio/lan-transfer/backup.zip'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, validPath)
-
-      expect(result).toBe(true)
-      expect(fs.remove).toHaveBeenCalledWith(normalizePath(validPath))
-      expect(mockLogger.info).toHaveBeenCalledWith(expect.stringContaining('Deleted temp backup'))
-    })
-
-    it('should delete file in nested subdirectory', async () => {
-      vi.mocked(fs.pathExists).mockResolvedValue(true as never)
-      vi.mocked(fs.remove).mockResolvedValue(undefined as never)
-
-      const nestedPath = '/tmp/cherry-studio/lan-transfer/sub/dir/file.zip'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, nestedPath)
-
-      expect(result).toBe(true)
-      expect(fs.remove).toHaveBeenCalledWith(normalizePath(nestedPath))
-    })
-
-    it('should return false when file does not exist', async () => {
-      vi.mocked(fs.pathExists).mockResolvedValue(false as never)
-
-      const missingPath = '/tmp/cherry-studio/lan-transfer/missing.zip'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, missingPath)
-
-      expect(result).toBe(false)
-      expect(fs.remove).not.toHaveBeenCalled()
-    })
-  })
-
-  describe('Path Traversal Attacks', () => {
-    it('should block basic directory traversal attack (../../../../etc/passwd)', async () => {
-      const attackPath = '/tmp/cherry-studio/lan-transfer/../../../../etc/passwd'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, attackPath)
-
-      expect(result).toBe(false)
-      expect(fs.pathExists).not.toHaveBeenCalled()
-      expect(fs.remove).not.toHaveBeenCalled()
-      expect(mockLogger.warn).toHaveBeenCalledWith(expect.stringContaining('outside temp directory'))
-    })
-
-    it('should block absolute path escape (/etc/passwd)', async () => {
-      const attackPath = '/etc/passwd'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, attackPath)
-
-      expect(result).toBe(false)
-      expect(fs.remove).not.toHaveBeenCalled()
-      expect(mockLogger.warn).toHaveBeenCalled()
-    })
-
-    it('should block traversal with multiple slashes', async () => {
-      const attackPath = '/tmp/cherry-studio/lan-transfer/../../../etc/passwd'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, attackPath)
-
-      expect(result).toBe(false)
-      expect(fs.remove).not.toHaveBeenCalled()
-    })
-
-    it('should block relative path traversal from current directory', async () => {
-      const attackPath = '../../../etc/passwd'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, attackPath)
-
-      expect(result).toBe(false)
-      expect(fs.remove).not.toHaveBeenCalled()
-    })
-
-    it('should block traversal to parent directory', async () => {
-      const attackPath = '/tmp/cherry-studio/lan-transfer/../backup/secret.zip'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, attackPath)
-
-      expect(result).toBe(false)
-      expect(fs.remove).not.toHaveBeenCalled()
-    })
-  })
-
-  describe('Prefix Attacks', () => {
-    it('should block similar prefix attack (lan-transfer-evil)', async () => {
-      const attackPath = '/tmp/cherry-studio/lan-transfer-evil/file.zip'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, attackPath)
-
-      expect(result).toBe(false)
-      expect(fs.remove).not.toHaveBeenCalled()
-      expect(mockLogger.warn).toHaveBeenCalled()
-    })
-
-    it('should block path without separator (lan-transferx)', async () => {
-      const attackPath = '/tmp/cherry-studio/lan-transferx'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, attackPath)
-
-      expect(result).toBe(false)
-      expect(fs.remove).not.toHaveBeenCalled()
-    })
-
-    it('should block different temp directory prefix', async () => {
-      const attackPath = '/tmp-evil/cherry-studio/lan-transfer/file.zip'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, attackPath)
-
-      expect(result).toBe(false)
-      expect(fs.remove).not.toHaveBeenCalled()
-    })
-  })
-
-  describe('Error Handling', () => {
-    it('should return false and log error on permission denied', async () => {
-      vi.mocked(fs.pathExists).mockResolvedValue(true as never)
-      vi.mocked(fs.remove).mockRejectedValue(new Error('EACCES: permission denied') as never)
-
-      const validPath = '/tmp/cherry-studio/lan-transfer/file.zip'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, validPath)
-
-      expect(result).toBe(false)
-      expect(mockLogger.error).toHaveBeenCalledWith(expect.stringContaining('Failed to delete'), expect.any(Error))
-    })
-
-    it('should return false on fs.pathExists error', async () => {
-      vi.mocked(fs.pathExists).mockRejectedValue(new Error('ENOENT') as never)
-
-      const validPath = '/tmp/cherry-studio/lan-transfer/file.zip'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, validPath)
-
-      expect(result).toBe(false)
-      expect(mockLogger.error).toHaveBeenCalled()
-    })
-
-    it('should handle empty path string', async () => {
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, '')
-
-      expect(result).toBe(false)
-      expect(fs.remove).not.toHaveBeenCalled()
-    })
-  })
-
-  describe('Edge Cases', () => {
-    it('should allow deletion of the temp directory itself', async () => {
-      vi.mocked(fs.pathExists).mockResolvedValue(true as never)
-      vi.mocked(fs.remove).mockResolvedValue(undefined as never)
-
-      const tempDir = '/tmp/cherry-studio/lan-transfer'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, tempDir)
-
-      expect(result).toBe(true)
-      expect(fs.remove).toHaveBeenCalledWith(normalizePath(tempDir))
-    })
-
-    it('should handle path with trailing slash', async () => {
-      vi.mocked(fs.pathExists).mockResolvedValue(true as never)
-      vi.mocked(fs.remove).mockResolvedValue(undefined as never)
-
-      const pathWithSlash = '/tmp/cherry-studio/lan-transfer/sub/'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, pathWithSlash)
-
-      // path.normalize removes trailing slash
-      expect(result).toBe(true)
-    })
-
-    it('should handle file with special characters in name', async () => {
-      vi.mocked(fs.pathExists).mockResolvedValue(true as never)
-      vi.mocked(fs.remove).mockResolvedValue(undefined as never)
-
-      const specialPath = '/tmp/cherry-studio/lan-transfer/file with spaces & (special).zip'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, specialPath)
-
-      expect(result).toBe(true)
-      expect(fs.remove).toHaveBeenCalled()
-    })
-
-    it('should handle path with double slashes', async () => {
-      vi.mocked(fs.pathExists).mockResolvedValue(true as never)
-      vi.mocked(fs.remove).mockResolvedValue(undefined as never)
-
-      const doubleSlashPath = '/tmp/cherry-studio//lan-transfer//file.zip'
-      const result = await backupManager.deleteLanTransferBackup({} as Electron.IpcMainInvokeEvent, doubleSlashPath)
-
-      // path.normalize handles double slashes
-      expect(result).toBe(true)
-    })
-  })
-})
