+++ src/renderer/utils/backup.ts
@@ -1,5 +1,12 @@
 import i18n from '@renderer/i18n/resolver'
-import { BACKUP_ACTIVE_WRITERS_ERROR_CODE } from '@shared/types/backup'
+
+/**
+ * Marker main puts in the error message when a backup is refused because data
+ * writers are still active. Renderer-only: nothing in main reads it back, so it
+ * stays out of `@shared` (shared-layer-architecture.md — cross-process is the
+ * entry gate, not a description).
+ */
+export const BACKUP_ACTIVE_WRITERS_ERROR_CODE = 'BACKUP_ACTIVE_WRITERS'
 
 type BackupErrorFallbackKey =
   | 'error.backup.file_format'
