+++ src/shared/IpcChannel.ts
@@ -81,8 +81,6 @@ export enum IpcChannel {
   File_RunSweep = 'file:runSweep',
 
   // backup
-  Backup_CreateLanTransferBackup = 'backup:createLanTransferBackup',
-  Backup_DeleteLanTransferBackup = 'backup:deleteLanTransferBackup',
 
   // zip
   Zip_Decompress = 'zip:decompress',
@@ -135,7 +133,8 @@ export enum IpcChannel {
   LanTransfer_Disconnect = 'lan-transfer:disconnect',
   LanTransfer_ClientEvent = 'lan-transfer:client-event',
   LanTransfer_SendFile = 'lan-transfer:send-file',
-  LanTransfer_CancelTransfer = 'lan-transfer:cancel-transfer'
+  LanTransfer_CancelTransfer = 'lan-transfer:cancel-transfer',
+  LanTransfer_DeleteFile = 'lan-transfer:delete-file'
 
   // ──────────────────────────────────────────────────────────────
   // TODO(v2): the following IPC channels are still referenced via
